\name{turboMEDUSA-internal}
\alias{calculateModelFit}
\alias{calculateModelFitSummary}
\alias{calculateModelWeights}
\alias{descendantsCutAtStem}
\alias{descendantsCutAtStem.idx}
\alias{descendantsCutAtNode}
\alias{descendantsCutAtNode.idx}
\alias{extractCladeFromTree}
\alias{findMrca}
\alias{getMaxModelLimit}
\alias{getNodeNumber}
\alias{getNumTips}
\alias{getOptimalModelFlavour}
\alias{getThreshold}
\alias{makeCacheMedusa}
\alias{makeLikMedusaPart}
\alias{medusaFitClade}
\alias{medusaFitUserNode}
\alias{medusaMLBaseModel}
\alias{medusaMLFitPartition}
\alias{medusaMLFitBase}
\alias{medusaMLPrefit}
\alias{medusaMLPrefitNode}
\alias{medusaMLPrefitStem}
\alias{medusaMLUpdate}
\alias{medusaSplit}
\alias{plotModelFit}
\alias{plotNN}
\alias{prepareData}
\alias{runMultiMEDUSA}
\alias{medusaSplitNode}
\alias{medusaSplitStem}

\title{Turbo-MEDUSA Internal Functions}
\description{\code{turboMEDUSA-internal} are either not typically called by the user or are currently undocumented}

\details{
This is an internal \pkg{turboMEDUSA} function, not intended to be called directly by the user.
}

\references{
Alfaro, ME, F Santini, C Brock, H Alamillo, A Dornburg, DL Rabosky, G Carnevale, and LJ Harmon. 2009. Nine exceptional radiations plus high turnover explain species diversity in jawed vertebrates. \emph{Proceedings of the National Academy of Sciences} \bold{106}: 13410-13414.
}

\author{
Joseph W. Brown, Richard G. FitzJohn, Michael E. Alfaro, and Luke J. Harmon.

Maintainer: Joseph W. Brown <josephwb@uidaho.edu>}