\name{turboMEDUSA-internal}
\alias{ancestors}
\alias{ancestors.idx}
\alias{calculate.model.fit}
\alias{calculate.model.fit.summary}
\alias{calculate.model.weights}
\alias{get.max.model.limit}
\alias{get.num.tips}
\alias{get.threshold}
\alias{make.cache.medusa}
\alias{make.lik.medusa.part}
\alias{medusa.ml.fit.partition}
\alias{medusa.ml.initial}
\alias{medusa.ml.prefit}
\alias{medusa.ml.update}
\alias{medusa.split}
\alias{plotModelFit}
\alias{prune.tree.merge.data}
\alias{runMultiMEDUSA}

\title{Turbo-MEDUSA Internal Functions}
\description{\code{turboMEDUSA-internal} are either not typically called by the user or are currently undocumented}

\details{
This is an internal \pkg{turboMEDUSA} function, not intended to be called directly by the user.
}

\references{
Alfaro, ME, F Santini, C Brock, H Alamillo, A Dornburg, DL Rabosky, G Carnevale, and LJ Harmon. 2009. Nine exceptional radiations plus high turnover explain species diversity in jawed vertebrates. \emph{Proceedings of the National Academy of Sciences} \bold{106}: 13410-13414.
}

\author{
Luke J. Harmon, Daniel L. Rabosky, Richard G. FitzJohn, and Joseph W. Brown

Maintainer: Joseph W. Brown <josephwb@uidaho.edu>}